Fixed Server (Game/Lobby) errors

+++ Network/Include/dr_packet.h
@@ -55,7 +55,7 @@ class DR_NETWORK_EXPORT Packet
     *this << strLen;
 
     for (auto& wc : data) {
-      *this << static_cast<UInt32>(wc);
+      *this << static_cast<UInt16>(wc);
     }
 
     return *this;
@@ -114,16 +114,16 @@ class DR_NETWORK_EXPORT Packet
 
     data.resize(size);
 
-    size *= sizeof(UInt32);
+    size *= sizeof(UInt16);
     
     if (size > 0) {
       Int8* buffer;
       getDataPtr(buffer, size);
       if (buffer) {
         for (auto& wc : data) {
-          UInt32 v = *reinterpret_cast<UInt32*>(buffer);
+          UInt16 v = *reinterpret_cast<UInt16*>(buffer);
           wc = static_cast<WString::value_type>(v);
-          buffer += sizeof(UInt32);
+          buffer += sizeof(UInt16);
         }
         //removeData(size);
       }
